Added more comments and renamed files
Added more comments and renamed files
